requirements.txt
flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
apscheduler==3.10.4
feedparser==6.0.10
python-dateutil==2.8.2
